feat: Split code and add account deletion tests
