+++ src/controllers/account-groups/delete-group.ts
@@ -1,6 +1,5 @@
 import { z } from 'zod';
 import type { CustomResponse } from '@common/types';
-import { API_RESPONSE_STATUS } from 'shared-types';
 import { recordId } from '@common/lib/zod/custom-types';
 import { errorHandler } from '@controllers/helpers';
 import * as accountGroupService from '@services/account-groups/account-groups.service';
@@ -10,14 +9,10 @@ export const deleteAccountGroup = async (req, res: CustomResponse) => {
     const { id: userId } = req.user;
     const { groupId }: DeleteAccountGroupParams['params'] = req.validated.params;
 
-    const deletedCount = await accountGroupService.deleteAccountGroup({
+    await accountGroupService.deleteAccountGroup({
       groupId,
       userId,
     });
-
-    if (deletedCount === 0) {
-      return res.status(404).json({ status: API_RESPONSE_STATUS.error });
-    }
 
     return res.status(204).send();
   } catch (err) {
