+++ src/services/account-groups/delete-account-group.e2e.ts
@@ -1,0 +1,24 @@
+import { expect } from '@jest/globals';
+import * as helpers from '@tests/helpers';
+
+describe('Delete account group', () => {
+  it('successfully deletes record', async () => {
+    const group = await helpers.createAccountGroup({
+      name: 'test',
+      raw: true,
+    });
+
+    const result = await helpers.deleteAccountGroup({
+      groupId: group.id,
+    });
+
+    expect(result.statusCode).toBe(204);
+  });
+  it('returns successful response for non-existing record deletion', async () => {
+    const result = await helpers.deleteAccountGroup({
+      groupId: 99999,
+    });
+
+    expect(result.statusCode).toBe(204);
+  });
+});
